move diffview to own folder. next step put all widgets in own file
